+++ src/play/game/storyboard_timeline.ts
@@ -167,285 +167,3 @@ export class StoryboardLayerTimeline extends Container {
     }
   }
 }
-
-const BACKTRACK_DEFAULT_VALUE_CLASSES = [
-  ["S", "V"],
-  ["C"],
-  ["R"],
-  ["C"],
-  ["P"],
-  ["F"],
-] as readonly CommandType[][];
-
-abstract class StoryboardRendererBase<T extends StoryboardSprite>
-  extends Sprite
-  implements IUpdatable
-{
-  private commandTimeline: Timeline<Command>;
-  private scalePositive = new Vector2(1, 1);
-  private isParameterActive: Record<ParameterType, boolean> = {
-    "": false,
-    H: false,
-    A: false,
-    V: false,
-  };
-
-  protected object: T;
-
-  constructor(object: T) {
-    super();
-    this.object = object;
-    this.position.copyFrom(object.startPosition);
-    this.anchor.copyFrom(ORIGIN_MAP.get(object.origin)!);
-
-    const timelineCommands = this.getTimelineCommands();
-
-    // TODO: Triggers
-    this.commandTimeline = new Timeline(
-      timelineCommands.map(this.createElement),
-      () => {},
-      this.updateCommand,
-      this.finalizeCommand,
-      false
-    );
-    this.tint = utils.rgb2hex([
-      STORYBOARD_BRIGHTNESS,
-      STORYBOARD_BRIGHTNESS,
-      STORYBOARD_BRIGHTNESS,
-    ]);
-
-    // Setup default values
-    for (const classTypes of BACKTRACK_DEFAULT_VALUE_CLASSES) {
-      for (const command of timelineCommands) {
-        if (classTypes.includes(command.type)) {
-          this.updateCommand(command, 0);
-          break;
-        }
-      }
-    }
-
-    // Setup default position
-    let xPosSet = false;
-    let yPosSet = false;
-    for (const command of timelineCommands) {
-      if (command.type === CommandType.Movement) {
-        if (!xPosSet) {
-          this.x = command.startValue.x;
-        }
-        if (!yPosSet) {
-          this.y = command.startValue.y;
-        }
-        break;
-      } else if (command.type === CommandType.MovementX) {
-        if (!xPosSet) {
-          this.x = command.startValue;
-          xPosSet = true;
-        }
-      } else if (command.type === CommandType.MovementY) {
-        if (!yPosSet) {
-          this.y = command.startValue;
-          yPosSet = true;
-        }
-      }
-
-      if (xPosSet && yPosSet) {
-        break;
-      }
-    }
-  }
-
-  private createElement = (command: Command): TimelineElement<Command> => ({
-    startTimeMs: command.startTime,
-    endTimeMs: command.endTime,
-    build: () => command,
-  });
-
-  private updateCommand = (command: Command, timeMs: number) => {
-    const { startTime, endTime, easing } = command;
-
-    if (endTime > startTime) {
-      const p = (timeMs - startTime) / (endTime - startTime);
-      const easingFn = EasingFunctions.getEasingFn(easing);
-
-      this.applyCommand(command, easingFn(p));
-    } else {
-      this.applyCommand(command, timeMs >= startTime ? 1 : 0);
-    }
-  };
-
-  private finalizeCommand = (command: Command) => {
-    this.applyCommand(command, 1);
-  };
-
-  private applyCommand(command: Command, p: number) {
-    const { startTime, endTime, startValue, endValue, parameter } = command;
-
-    switch (command.type) {
-      case CommandType.Parameter: {
-        // BlendingCommand, HorizontalFlipCommand, and VerticalFlipCommand
-        // Active for the duration of the command
-        // If endTime == startTime, it's applied forever (?)
-        this.isParameterActive[parameter] = p < 1 || startTime === endTime;
-        return;
-      }
-      case CommandType.Color: {
-        const { red, green, blue } = lerpRGB(p, startValue, endValue);
-
-        this.tint = utils.rgb2hex([
-          (red / 255) * STORYBOARD_BRIGHTNESS,
-          (green / 255) * STORYBOARD_BRIGHTNESS,
-          (blue / 255) * STORYBOARD_BRIGHTNESS,
-        ]);
-
-        return;
-      }
-      case CommandType.Fade: {
-        this.alpha = lerp(p, startValue, endValue);
-        return;
-      }
-      case CommandType.Movement: {
-        this.x = lerp(p, startValue.x, endValue.x);
-        this.y = lerp(p, startValue.y, endValue.y);
-        return;
-      }
-      case CommandType.MovementX: {
-        this.x = lerp(p, startValue, endValue);
-        return;
-      }
-      case CommandType.MovementY: {
-        this.y = lerp(p, startValue, endValue);
-        return;
-      }
-      case CommandType.Rotation: {
-        this.rotation = lerp(p, startValue, endValue);
-        return;
-      }
-      case CommandType.Scale: {
-        const scale = lerp(p, startValue, endValue);
-
-        this.scalePositive.x = scale;
-        this.scalePositive.y = scale;
-        return;
-      }
-      case CommandType.VectorScale: {
-        this.scalePositive.x = lerp(p, startValue.x, endValue.x);
-        this.scalePositive.y = lerp(p, startValue.y, endValue.y);
-        return;
-      }
-    }
-
-    console.warn("Unknown command", command);
-  }
-
-  update(timeMs: number): void {
-    this.commandTimeline.update(timeMs);
-
-    // Apply parameters
-    this.scale.x = this.isParameterActive.H
-      ? -this.scalePositive.x
-      : this.scalePositive.x;
-    this.scale.y = this.isParameterActive.V
-      ? -this.scalePositive.y
-      : this.scalePositive.y;
-    this.blendMode = this.isParameterActive.A
-      ? BLEND_MODES.ADD
-      : BLEND_MODES.NORMAL;
-  }
-
-  private getTimelineCommands(): Command[] {
-    // Combine all commands, loops and triggers into one action timeline.
-    // TODO: Triggers
-    const timelineCommands = [
-      ...this.object.timelineGroup.commands,
-      ...this.object.loops.flatMap(this.unrollLoopCommand),
-    ];
-
-    return timelineCommands.sort((a, b) => a.startTime - b.startTime);
-  }
-
-  private unrollLoopCommand(loop: CommandLoop): Command[] {
-    const baseCommands = loop.commands;
-
-    if (baseCommands.length == 0) {
-      console.warn("Loop has no valid commands");
-      return [];
-    }
-
-    const unrolledCommands: Command[][] = [];
-
-    for (let i = 0; i < loop.totalIterations; i++) {
-      const iterationStartTime = loop.loopStartTime + i * loop.commandsDuration;
-
-      const cloned = baseCommands.map((command) => {
-        return new Command({
-          ...command,
-          startTime: command.startTime + iterationStartTime,
-          endTime: command.endTime + iterationStartTime,
-        });
-      });
-
-      unrolledCommands.push(cloned);
-    }
-
-    return unrolledCommands.flat();
-  }
-}
-
-class StoryboardSpriteRenderer extends StoryboardRendererBase<StoryboardSprite> {
-  public constructor(
-    storyboardResources: Map<string, Texture>,
-    object: StoryboardSprite
-  ) {
-    super(object);
-    this.texture = storyboardResources.get(object.filePath) ?? Texture.EMPTY;
-    if (this.texture == Texture.EMPTY) {
-      console.warn("Sprite has no texture");
-    }
-  }
-}
-
-class StoryboardAnimationRenderer extends StoryboardRendererBase<StoryboardAnimation> {
-  private frames: Texture[];
-  protected startTime: number;
-
-  public constructor(
-    storyboardResources: Map<string, Texture>,
-    object: StoryboardAnimation
-  ) {
-    super(object);
-
-    /*
-    TODO: This is incorrect, but works for 99% of maps
-    The only map I've seen it affect is https://osu-js.pages.dev/play/?1006822
-
-    I believe osu!stable calculates from the first *visible* frame.
-    This is hard to calculate because it's very dependent on the internal logic of osu!stable
-    and of the interactions between parameters, color, scale, alpha, and maybe even position?
-
-    osu!lazer currently (2020-8-5) uses this flawed/simplified logic:
-    */
-    this.startTime = object.startTime;
-
-    this.frames = getAllFramePaths(object).map(
-      (path) => storyboardResources.get(path) ?? Texture.EMPTY
-    );
-    if (this.frames.findIndex((x) => x == Texture.EMPTY) >= 0) {
-      console.warn("Animation missing frames");
-    }
-    this.texture = this.frames[0];
-  }
-
-  update(timeMs: number): void {
-    super.update(timeMs);
-
-    let frameNumber = Math.floor(
-      (timeMs - this.startTime) / this.object.frameDelay
-    );
-    if (this.object.loopType === LoopType.LoopForever) {
-      frameNumber = frameNumber % this.frames.length;
-    } else {
-      frameNumber = Math.min(frameNumber, this.frames.length - 1);
-    }
-    this.texture = this.frames[frameNumber];
-  }
-}
