+++ src/play/render/standard/components/circle_triangles.frag
