+++ src/play/render/standard/components/slider_path.frag
