+++ src/play/render/standard/components/slider_path.ts
