+++ src/play/render/standard/components/slider_reverse_arrow.ts
