+++ src/play/render/standard/components/slider_tick.ts
