+++ src/play/render/standard/cursor.ts
