+++ src/play/render/standard/cursor_autoplay.ts
