+++ src/play/render/standard/slider.ts
